replaced school metrics with voluteer opportunities added VolunteerList component fixes #261
